refactor(msg): consolidate pack_call/unpack_call into msg::callsign28
Bottom-up sweep continuation of the code-sharing audit (#290-296),
this time at the msg/ layer instead of engine/'s DSP layer per the
user's request to check core/fec/msg/protocol for gaps.

fec/ and the trait layer (engine/protocol.rs's per-protocol impls)
came back clean on inspection: fec::ldpc/qra already correctly share
their bp/osd/QraCode engines with ldpc240_101/ldpc_128_90/
qra15_65_64 as thin per-code-table wrappers (the same shape this
session already extracted for interleave/spectrogram), fec::conv
already houses ConvFano/ConvFano232 as wrappers over one shared fano
module, and the per-protocol ModulationParams/FrameLayout/Protocol
impls are const data that's supposed to differ by design. Grid
packing (jt72::pack_grid4_plain vs wspr::pack_grid4_power) looked
promising on the pack side (same core arithmetic) but the unpack
sides use genuinely different algorithms — jt72 inverts the packed
integer directly, wspr round-trips through lat/long degrees the way
wsprd's own C source does — so left alone as two faithful ports of
two different upstream routines, not a duplicate.

msg/jt72.rs and msg/wspr.rs's pack_call/unpack_call were the real
find: byte-identical base-37/36/10/27³ callsign encoding (same
NBASE=262_177_560, same CALL_ALPHA/CHAR37 table, same digit-position
shift rule), independently ported from WSJT-X's packjt.f90::packcall/
unpackcall — confirmed via grep that wsprcode/wspr_old_subs.f90 calls
that exact routine too, not a WSPR-specific one. JT65 already shares
jt72.rs with JT9 (both route through msg::Jt72Codec), so this was
WSPR's own third independent copy of the same upstream subroutine.

Extracted to msg::callsign28::{pack_call28, unpack_call28}. The one
real behavioural axis (which characters a slot accepts) stays a
closure parameter rather than folded in: jt72's nchar accepts a-z as
well as A-Z (dead in practice — callers already normalize case, and
nchar(lowercase) == nchar(uppercase) by construction, verified before
relying on it), wspr's callsign_char_code is uppercase-only. Special
CQ/QRZ/DE tokens have no WSPR analogue (WSPR Type-1 messages are
always callsign+grid+power) so that layer stays in jt72 alone, above
the shared core.

One incidental correctness fix, not just a refactor: WSPR's own
pack_call was missing three slot-alphabet checks that WSJT-X's
packcall.f90 does perform (slot 1 must not be space, slot 2 must be a
digit, slots 3-5 must not be digits) — confirmed against
packjt.f90:97-116's n1..n6 validation. The shared core carries the
full check (it was already correct in jt72's copy), so WSPR gains it
too. No existing WSPR test exercised the gap (both golden-corpus
callsigns are well-formed), and this can only make previously-
malformed input correctly return None, never change output on valid
input — verified via the full golden/fixed-point suite below.

Verified: full merge-gate suite (tier A+B, full+internal-testing,
428+ tests) green, JT9/JT65/WSPR golden tests pass under both full
and fixed-point features, isolated --features {jt9,jt65,wspr} builds
clean, fmt/clippy/doc clean. No ratchet entry added — this
consolidation is entirely internal to msg/ (an already-shared
directory), the same category as the pre-existing fec::ldpc/qra
internal sharing, not a per-protocol-adoption question the ratchet
tracks.
